In this project the user can retrieve the information according to his/her own preferences
In this project we have implemented an
social adaptive navigation support for discussion
forum posts that comprises navigable recommender
system which present users with results based on
their queries and their profiles. For this project, we
have considered only the java based posts obtained
from the stackoverflow website. We have
implemented different visualization techniques like
Sankey visualization that are helpful to the user for
navigation and learning. In addition to visualization
we have performed nearest neighbor computation
based on pearson’s correlation coefficient.